KEA-83 split Header component into a Layout component with Header and NavMenu components
